Separate 'package agent' into 'package streaming' and 'package polling'
