+++ probe/ebpf/ebpf.go
@@ -1,5 +1,3 @@
-// +build linux
-
 package ebpf
 
 import (
